refactor(test): update all test files to refer to test-utils
