+++ package.json
@@ -58,6 +58,7 @@
     "@storybook/testing-library": "^0.0.13",
     "@testing-library/react-hooks": "^8.0.1",
     "accessibility-checker": "^3.1.47",
+    "babel-jest": "^29.5.0",
     "babel-plugin-named-exports-order": "^0.0.2",
     "eslint": "^8.38.0",
     "eslint-plugin-react": "^7.32.2",
