+++ src/components/Banner/Banner.test.a11y.js
@@ -2,7 +2,7 @@
   This is a a11y testing file for the Banner connected component
  */
 import React from 'react';
-import { render } from '../../../test-utils';
+import { render } from '../../../config/jest/test-utils';
 import Banner from './Banner';
 
 describe('Banner a11y', () => {
