+++ src/components/Banner/Banner.test.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import { render, screen } from '../../../test-utils';
+import { render, screen } from '../../../config/jest/test-utils';
 import Banner from './Banner';
 
 describe('Banner', () => {
