+++ src/components/EnhancedDataTable/EnhancedDataTable.test.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import { render, screen, waitFor } from '../../../test-utils';
+import { render, screen, waitFor } from '../../../config/jest/test-utils';
 import userEvent from '@testing-library/user-event';
 import useResizeObserver from 'use-resize-observer/polyfilled';
 import { OverflowMenuItem } from '@carbon/react';
