+++ src/components/IsometricIcon/IsometricIcon.test.a11y.js
@@ -2,7 +2,7 @@
   This is a a11y testing file for the IsometricIcon connected component
  */
 import React from 'react';
-import { render } from '../../../test-utils';
+import { render } from '../../../config/jest/test-utils';
 import IsometricIcon from './IsometricIcon';
 
 describe('IsometricIcon a11y', () => {
