+++ src/components/MediaGallery/MediaGallery.test.a11y.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import { act, render } from '../../../test-utils';
+import { act, render } from '../../../config/jest/test-utils';
 import MediaGallery from './MediaGallery';
 
 const props = {
