+++ src/components/OrderSummaryV2/OrderSummaryV2.test.js
@@ -5,7 +5,7 @@ import {
   Link,
   InlineNotification,
 } from '@carbon/react';
-import { render, screen, waitFor } from '../../../test-utils';
+import { render, screen, waitFor } from '../../../config/jest/test-utils';
 import userEvent from '@testing-library/user-event';
 import OrderSummaryV2 from './OrderSummaryV2';
