+++ src/components/PageHeader/PageHeader.test.a11y.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import { render } from '../../../test-utils';
+import { render } from '../../../config/jest/test-utils';
 // Page Header Components
 import PageHeader from './PageHeader';
