+++ src/components/PageHeader/PageHeader.test.js
@@ -1,7 +1,7 @@
 import '@testing-library/jest-dom'
 
 import React from 'react';
-import { render } from '../../../test-utils';
+import { render } from '../../../config/jest/test-utils';
 import PageHeader from './PageHeader';
 import PageHeaderSkeleton from './skeleton';
