+++ src/components/SideNav/SideNav.test.a11y.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import { render, fireEvent } from '../../../test-utils';
+import { render, fireEvent } from '../../../config/jest/test-utils';
 import SideNav from './SideNav';
 
 const itemsWithSubMenu = [
