+++ src/components/SideNav/SideNav.test.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import { render, screen, fireEvent } from '../../../test-utils';
+import { render, screen, fireEvent } from '../../../config/jest/test-utils';
 import userEvent from '@testing-library/user-event';
 import SideNav from './SideNav';
