+++ src/components/Translate/Translate.test.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import { render } from '../../../test-utils';
+import { render } from '../../../config/jest/test-utils';
 import Translate from './Translate';
 
 describe('Translate', () => {
